requirements.txt
asgiref==3.7.2
astroid==3.0.2
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
click==8.1.7
coverage==7.5.1
decouple==0.0.7
dill==0.3.8
Django==5.0.3
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==22.5.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.1.0
pluggy==1.5.0
PyJWT==2.8.0
pylint==3.0.3
PySocks==1.7.1
pytest==8.2.0
pytest-django==4.8.0
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.1
wsproto==1.2.0
